+++ airflow-core/src/airflow/utils/dot_renderer.py
@@ -25,9 +25,9 @@ from typing import TYPE_CHECKING, Any
 
 from airflow.exceptions import AirflowException
 from airflow.sdk import BaseOperator
+from airflow.sdk.definitions.edges import dag_edges
 from airflow.sdk.definitions.mappedoperator import MappedOperator
 from airflow.serialization.serialized_objects import SerializedBaseOperator
-from airflow.utils.dag_edges import dag_edges
 from airflow.utils.state import State
 from airflow.utils.task_group import TaskGroup
